# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FleetSaaS school bus tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List all routes for this tenant
      responses:
        "200":
          description: Route list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Create a new route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "201":
          description: Created route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/{id}:
    patch:
      operationId: updateRoute
      tags: [routes]
      summary: Update route name or assigned driver/vehicle (fleet swap)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteUpdate"
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Delete a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Route:
      type: object
      required: [id, tenantId, name, isActive]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        driverId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        driverName:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
        departureTime:
          type: string
          description: "Base departure time string e.g. '06:00 AM'"
        avgSpeedKmh:
          type: integer
          description: Average fleet speed in km/h for ETA calculation
    RouteInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        driverId:
          type: integer
        vehicleId:
          type: integer
        departureTime:
          type: string
        avgSpeedKmh:
          type: integer
    RouteUpdate:
      type: object
      properties:
        name:
          type: string
        driverId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        departureTime:
          type: string
        avgSpeedKmh:
          type: integer
...